reticulated/pyproject.toml
[tool.poetry]
name = "reticulated"
version = "0.1.0"
description = "Python monobuild system"
authors = ["Joshua Cannon <joshdcannon@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.2"
click = "^7.1.2"
pluggy = "^0.13.1"
toml = "^0.10.2"
packaging = "^20.9"
attrs = "^21.2.0"
cached-property = "^1.5.2"

[tool.poetry.dev-dependencies]
black = "^21.5b0"
flake8 = "^3.9.1"
pytest = "^6.2.4"
flake8-import-order = "^0.18.1"
flake8-docstrings = "^1.6.0"
Sphinx = "^3.5.4"
pytest-mock = "^3.6.1"
sphinx-rtd-theme = "^0.5.2"

[tool.poetry.scripts]
reticulated = 'reticulated.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
